Restructure and rework
Use buttons instead of an input field, to clarify that this is a learning example
not a real-world project (without a proper use case...)

Add placeholder space to hold fetched images that would be populated using AJAX or
WebSockets, once the image has been fetched. This isn't implemented yet.

Experiment with non-working HTTP schenanigans, and with base class settings for
Celery Tasks.

Probably storing the project with this commit and starting from scratch with a different
idea. :P