Use fake job board
indeed.com has tightened their bot protection against web scraping, which is why requests to their site as they are described in this course return 403 Forbidden status codes.

I've attempted to circumvent this using fake headers (something that would be explainable in an intro course) but no luck, 403 prevails.

I've previously [reworked the written tutorial](https://realpython.com/beautiful-soup-web-scraper-python/#step-1-inspect-your-data-source) to use a self-hosted [fake job board](https://realpython.github.io/fake-jobs/) that I set up just for the purpose of the tutorial.

As a quick fix for the video course, I added an explanatory lesson to the video coure and reworked the Jupyter notebooks.

The information and processes that I explain in the rest of the course are still valid and a good introduction for how to approach scraping a static website.