Restructures project to work with /discover page
Bandcamp removed the "Discover" section from their main page
since we wrote this code. Now, the tracks are only available at the
dedicated /discover URL.

I restructured the code to target the /discover site instead, which
required a couple of changes. Still, the existing POM structure was
helpful :)

I also (re)introduced a new "pause" option, because that's a bit easier
in this new structure.

Finally, I only allow loading more songs once, which gives a total of
120 songs to pick from with the default screen setting. Attempting to
load more resulted in errors, I think because they'd be outside of the
viewport and I didn't want to expand the code more and introduce
scrolling more of them into view. Also, not sure whether that'd take
earlier ones out of the viewport (etc) so I just decided not to open
that can of worms for this "Intro to Selenium" tutorial. LMK if you
disagree, otherwise of course you can tackle that in your longer one
that builds on top of this one :)